Bump `windows(-core)` crate range to `0.57-0.59`
The generated code is no longer compatible with `0.56` and requires
`0.57`. Also, `windows-bindgen 0.59` created a slightly different
layout.

https://github.com/microsoft/windows-rs/releases/tag/0.61.0
