+++ Cargo.toml
@@ -13,12 +13,11 @@ include = ["src", "LICENSE"]
 anyhow = "1.0.3"
 libloading = "0.8"
 # Autogenerated bindings reference `::windows_core` instead of `::windows::core`
-windows-core = ">=0.56, <=0.58"
+windows-core = ">=0.57, <=0.59"
 
 [dependencies.windows]
-version = ">=0.56, <=0.58"
+version = ">=0.57, <=0.59"
 features = [
-    "implement",
     "Win32_Foundation",
     "Win32_Graphics_Direct3D12",
     "Win32_Graphics_Dxgi_Common",
